// src/typings/api/trade.d.ts
export type AdjustCoinGroupType = {
  id: string
  /**
   * 合约组名称
   */
  name: string
  /**
   * 是否自动追加保证金
   */
  isAutoAdd: string
  /**
   * 已实现盈亏
   */
  realizedProfit: string
  /**
   * 计价币
   */
  baseCoin: string
  /**
   * 未实现盈亏
   */
  unrealizedProfit: string
}